ENH: Add bsplines.py and test_bsplines.py
This is experimental work in progress. Currently there the bsplines.py
file contains a Tck class to hold validated values of the knot points,
coefficients, dtype, and degree. The aim is that having such a class
allows calling Cythonized C routines without having to run a bunch of
checks to avoid segfaults and such and to provide a base that can be
subclassed. The current class is only 1-D and will need modifications.

There are also subroutines:

    bsplvander  -- psuedo Vandermonde matrices for b-splines
    bsplval     -- evaluate a b-spline
    bsplderiv   -- differentiate a b-spline

They are very slow at this point due to a lot of indexing. It is intended
that they be converted to useful C at some future time when they have been
tested and the API somewhat finalized.

The tests are pretty basic, but they serve to avoid obvious errors.