ENH: Add Tck support for tensor products and multiple dimension values.
The setup is that tensored b-splines map a polytope in E^m into the
vector space R^n. Note that the vector space may be an array with
multiple dimensions, i.e., an arrays with shape (2, 3) are taken to
be a vector space of dimension 6.

To support this functionality, the degrees of the b-splines that will be
tensored are kept in a list and the corresponding knot points in the
form of ndarrays are kept in another list. The coeffients are kept in a
single ndarray. The first m dimensions of the coefficient array
correspond to the m b-splines that will be tensored, and the remaining
dimensions are for the range space. As mentioned above, the range space
can be multi-dimensional and as a vector space its dimension is the
product of the remaining dimensions.

This can be illustrated in two tensored dimensions as

    f(x,y) = \sum_{j,m} c[j,m,...]*N_{i,j}(x)*N_{l,m}(y)

where f(x,y) is the b-spline value at (x, y), c is the coefficient array
and N_{i,j} is the j'th b-spline of degree i.

degrees into a list of degrees and knot points into a list of knot
points. The coefficient array has the tensor dimensions in the first
shape elements, and the vector