ENH: Add addition and subtraction operators to Tck object.
There are two cases here:

    1) Both objects are Tck objects
    2) One of the object is array_like

In the first case, both Tck objects must have the same type, knot
points. If n is the number of leading indices in the coefficient shapes
that index over the domain, then ones are inserted into the shape of the
smaller coefficient array starting at the n'th position until ther
arrays have the same number of dimensions. The effect of this is that
the vector values of the splines broadcast in the same way as standard
numpy ndarrays.

In the second case, the array_like object is first converted to an array
of the same dtype as the existing coefficient array and, if it has equal
or fewer dimensions than the range dimensions of the Tck object and
broadcasts to those coefficients, then it is added/subtracted. The
reason is isn't simply added without the dimension check is that some of
the Tck domain dimensions can be legitimately 1 and we do not want the
domain dimensions to change. Effectively, the array being added is a
scalar.