ENH: Add multiplication/division operators to Tck objects.
The valid types for these operations are 'array scalars', Tck objects do
not support multiplication/division with other Tck object. An array
scalar in this context is an array_like object that broadcasts over
coefficient array indices devoted to the domain dimensions. For example,
if two splines are tensored to give two spacial dimensions, then the
first two indexes of the coefficient array are the domain indexes and
the broadcast array must have at least two fewer indexes then the
coefficient array. We are essentially using the same array (vector) at
every knot point of the domain, it is like a constant function.