ENH: Make functions and tests tensor/multidimensional compatible.
The changes to Tck to support tensor product splines with vector
values made Tck incompatible with the previous functions in bsplines.py
that made use of the Tck type. The fixes those functions the test that go
with them.

The biggest change that may be notices by the user is that when a bspline
is evaluated it always returns a Tck object. When the tensored b-splines
are completely reduced the domain dimensionality becomes zero and the
results are in the coefficient array. It might be desireable to simply
return the results in that case, but that is for another day.

Tests pass, but this part of the work is incomplete and needs much more
testing.