ENH: Control access to some of the Tck internals.
The goal is to make sure that the consistency of the Tck data is
maintained. To that end the following policy is implemented:

    1) Methods that return the coeffient array return a view instead.
    2) Methods that return the knot points list return a shallow copy.
    3) Methods that return the domain list return a shallow copy.
    4) Methods that return the degrees list return a shallow copy.
    5) The knot point arrays are made not writeable.
    6) The domain list elements are tuples of numbers, hence immutable.
    7) All internal variables get the '__' prefix.

Note that the returned coefficient view can be used to change the
coefficients, and it can even be reshaped (bad idea), but the shape and
dtype of the internal array will remain fixed.