REF: Change argument order of BSpline.
The new order is almost the same as in the statsmodels BSpline class but
with degree instead of order. The new order is (t, k, c), where t are the
knot points, k are the degrees, and c is the coefficients.  The virtue of
this order is that c can have a default value of None, which makes for a
useful domain only class. Getting all the wanted functionality by, say,
inheriting from a domain class isn't that easy to do and this is pretty
easy.