REF: Break BSDomain out of BSpline and use informative names.
It is cleaner to have the class indicate the presence/absence of
coefficients rather than a None value. It is also neater since
comparisons for equality are easier to decide on for knot points than
for arrays.

The name of the attributes have been made more informative:

    t -> knots
    c -> coef
    k -> degrees
    ...

The bsplval function now returns an array when the domain dimension goes
to zero rather than a BSpline object. This is less likely to confuse and
bother users than needing to access the coefficient attribute.